requirements.txt
amqp==2.5.1
aniso8601==7.0.0
asgiref==3.3.1
autopep8==1.4.4
Babel==2.7.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.6.1.0
bleach==2.1.4
boto3==1.9.213
botocore==1.12.213
braintree==3.49.0
cachetools==3.1.1
cairocffi==1.0.2
CairoSVG==2.4.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-email-url==0.2.0
Django==3.1.3
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-cache-url==3.0.0
django-ckeditor==6.0.0
django-countries==5.4
django-elasticsearch-dsl==0.5.1
django-email-extras==0.3.4
django-filter==2.2.0
django-graphql-jwt==0.2.1
django-impersonate==1.4.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-measurement==3.0.0
django-mptt==0.10.0
django-phonenumber-field==2.4.0
django-recaptcha==1.5.0
django-redis==4.10.0
django-render-block==0.6
django-silk==2.0.0
django-storages==1.7.1
django-templated-email==2.3.0
django-tinymce==3.2.0
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
djangorestframework==3.10.3
docutils==0.15.2
draftjs-sanitizer==1.0.0
elasticsearch==6.3.1
elasticsearch-dsl==6.0.1
et-xmlfile==1.0.1
freezegun==0.3.12
future==0.17.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.18.0
google-i18n-address==2.3.5
google-measurement-protocol==1.0.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gprof2dot==2016.10.13
graphene==2.1.8
graphene-django==2.5.0
graphene-django-optimizer==0.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
html-to-draftjs==1.0.1
html5lib==1.0.1
idna==2.8
importlib-metadata==0.19
ipaddress==1.0.22
jdcal==1.4.1
Jinja2==2.10.1
jmespath==0.9.4
jsonfield==2.0.2
kombu==4.6.4
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
measurement==3.0.0
mpmath==1.1.0
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.3
phonenumberslite==8.10.17
Pillow==5.4.1
prices==1.0.0
promise==2.2.1
protobuf==3.9.1
psycopg2-binary==2.8.3
purl==1.5
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.8.0
python-gnupg==0.4.5
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
razorpay==1.2.0
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
Rx==1.6.1
s3transfer==0.2.1
sentry-sdk==0.11.1
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.3
sqlparse==0.3.0
stripe==2.35.1
sympy==1.4
tablib==0.13.0
text-unidecode==1.2
tinycss2==1.0.2
tqdm==4.34.0
typing==3.7.4
Unidecode==1.1.1
urllib3==1.25.3
vine==1.3.0
WeasyPrint==48
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.5.2
